feat: Refactor profile page and separate in components
